# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MNT Embark API — luxury tour company
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: All locations with country name for admin selects
      responses:
        "200":
          description: Location list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationSummary"
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create a new location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationInput"
      responses:
        "201":
          description: Created location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationSummary"
  /locations/{id}:
    patch:
      operationId: updateLocation
      tags: [locations]
      summary: Update a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdateInput"
      responses:
        "200":
          description: Updated location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationSummary"
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]

  # ── Countries ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LocationSummary:
      type: object
      properties:
        id:
          type: number
        slug:
          type: string
        name:
          type: string
        countryId:
          type: ["number", "null"]
        countryName:
          type: ["string", "null"]
        image:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        displayOrder:
          type: number
      required: [id, slug, name]
    LocationInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        slug:
          type: ["string", "null"]
        countryId:
          type: ["number", "null"]
        image:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        displayOrder:
          type: number
      required: [name]
    LocationUpdateInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: ["string", "null"]
        countryId:
          type: ["number", "null"]
        image:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        displayOrder:
          type: number
...